Rename features folder to components and create assets folder
